Create Number Reader
A program that takes a whole number (without a decimal point) between -999999999999 and +999999999999 and prints it in words on screen. For instance, if the number 9876 is given to the program, it prints "Nine thousand, eight hundred and seventy six" and for -100670011 it prints "Negative one hundred million, six hundred and seventy thousand, eleven".